backend/requirements.txt
# --- Backend & Server ---
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2

# --- AI & Computer Vision ---
ultralytics==8.3.0
opencv-python==4.10.0.84
numpy==1.26.4

# --- Deep Learning (Windows는 설치 방식 주의) ---
# 아래 torch 관련은 pip install -r 대신 별도 명령어를 권장할 수 있음

# NVIDIA 그래픽카드가 있는 경우 

# NVIDIA 공식 사이트에서 CUDA Toolkit(11.8 또는 12.1 권장)을 먼저 설치해야 합니다.
# 이후 터미널에서 아래 명령어로 PyTorch를 먼저 설치한 뒤 requirements.txt를 실행합니다.
# pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu121
# pip install -r requirements.txt

torch
torchvision
torchaudio

# --- Utilities ---
requests==2.32.3
python-dotenv==1.0.1
websockets==13.1